backend/requirements.txt
Django==5.0.6
mysqlclient
djangorestframework
PyMySQL==1.1.1
python-decouple
drf-yasg
gunicorn
django-cors-headers
